# rock-platform/rock/docs/swagger.yaml
definitions:
  api.UserDetailResp:
    properties:
      created_at:
        example: "2020-12-20 15:15:22"
        type: string
      email:
        example: admin_user@sensetime.com
        type: string
      id:
        description: type UserFullResp struct {
        example: 1
        type: integer
      login_block_until:
        example: "2020-12-20 15:15:22"
        type: string
      name:
        example: admin_user
        type: string
      password:
        example: '********'
        type: string
      role_created_at:
        example: "2020-12-20 15:15:22"
        type: string
      role_description:
        example: description for role
        type: string
      role_id:
        type: integer
      role_name:
        type: string
      role_updated_at:
        example: "2020-12-20 15:15:22"
        type: string
      role_version:
        example: 1
        type: integer
      salt:
        example: salt secret
        type: string
      token:
        example: user token
        type: string
      updated_at:
        example: "2020-12-20 15:15:22"
        type: string
    required:
    - role_description
    - role_version
    type: object
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        description: can not use LocalTime
        type: string
      email:
        type: string
      id:
        type: integer
      login_block_until:
        type: string
      login_retry_count:
        type: integer
      name:
        type: string
      password:
        type: string
      reset_secret:
        type: string
      role_id:
        description: Role            Role       `json:"role" gorm:"ForeignKey:RoleId;AssociationForeignKey:Id"`
          // use RoleId to ForeignKey
        type: integer
      salt:
        type: string
      secret_expired_at:
        type: string
      token:
        description: |-
          GitlabToken     string     `json:"gitlab_token"`
          DroneToken      string     `json:"drone_token"`
        type: string
      updated_at:
        type: string
      version:
        description: |-
          DeletedAt LocalTime `json:"deleted_at" gorm:"type:timestamp;default:null"`  // error
          DeletedAt LocalTime `json:"deleted_at" gorm:"type:timestamp null"`  // error
        type: integer
    type: object
  models.UserPagination:
    properties:
      items:
        items:
          $ref: '#/definitions/models.User'
        type: array
      page_num:
        type: integer
      page_size:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  utils.HTTPError:
    properties:
      error:
        example: error message here
        type: string
      error_code:
        example: 50000001
        type: integer
    type: object
...
  v1.CreateUserReq:
    properties:
      email:
        example: admin_user@sensetime.com
        type: string
      name:
        example: admin_user
        type: string
      password:
        example: '********'
        type: string
      role_id:
        description: RoleId   *RoleIdReq `json:"role_id" binding:"required"`  // 用顺义的这种定义，ctx.ShouldBind报错
        example: 1
        type: integer
    required:
    - email
    - name
    - password
    - role_id
    type: object
...
  v1.PaginateBriefUserResp:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.UserBriefResp'
        type: array
      page_num:
        example: 1
        type: integer
      pages:
        example: 1
        type: integer
      per_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    required:
    - items
    - page_num
    - pages
    - per_size
    - total
    type: object
...
  v1.UpdateUserPwdReq:
    properties:
      new_password:
        example: '********'
        type: string
      old_password:
        example: '********'
        type: string
    required:
    - new_password
    - old_password
    type: object
  v1.UserBriefResp:
    properties:
      created_at:
        example: "2018-10-09T14:57:23+08:00"
        type: string
      email:
        example: admin@sensetime.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: admin_role
        type: string
      role_id:
        example: 1
        type: integer
      updated_at:
        example: "2018-10-09T14:57:23+08:00"
        type: string
      version:
        example: 1
        type: integer
    required:
    - created_at
    - email
    - id
    - name
    - role_id
    - updated_at
    - version
    type: object
info:
  contact: {}
paths:
...
  /v1/roles/{id}/users:
    get:
      consumes:
      - application/json
      description: api for get all users by species role id
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Request page number
        in: query
        name: page_num
        required: true
        type: integer
      - default: 10
        description: Request page size
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: StatusOK
          schema:
            $ref: '#/definitions/v1.PaginateBriefUserResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get all users by species role id
      tags:
      - ROLE
  /v1/users:
    get:
      consumes:
      - application/json
      description: Api to get all users
      parameters:
      - default: 1
        description: Request page number
        in: query
        name: PageNum
        required: true
        type: integer
      - default: 10
        description: Request page size
        in: query
        name: PageSize
        required: true
        type: integer
      - description: 'Fuzzy Query(field: name)'
        in: query
        name: QueryField
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserPagination'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get users
      tags:
      - USER
    post:
      consumes:
      - application/json
      description: Api to create user
      parameters:
      - description: JSON type input body
        in: body
        name: input_body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserDetailResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Create user
      tags:
      - USER
  /v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Api to delete user with id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: StatusNotFound
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Delete user with id
      tags:
      - USER
    get:
      consumes:
      - application/json
      description: Api to get user with id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserDetailResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: StatusNotFound
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get user with id
      tags:
      - USER
  /v1/users/{id}/password:
    put:
      consumes:
      - application/json
      description: Api to update user password with id and old password
      parameters:
      - description: User ID
        in: path
        name: Id
        required: true
        type: integer
      - description: JSON body for update user info
        in: body
        name: update_body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateUserPwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserDetailResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: StatusNotFound
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Update user password with id and old password
      tags:
      - USER
swagger: "2.0"